Nginx config for graphite: enable CORS (e.g. for Grafana)
This commit must use some "tricks" to shape the Hiera data in a way that
it achieves the correct affect with the version of puppet-nginx we
currently use in Wirbelsturm.

Once we upgrade puppet-nginx to a newer version the nginx configuration
in Hiera should be clean again (e.g. because we can then use
`raw_append`, which is a new parameter in recent puppet-nginx versions).